requirements.txt
fastapi==0.127.1
uvicorn==0.40.0
pydantic==2.12.5
timezonefinder==8.2.0
tzdata==2025.3
pyswisseph==2.10.3.2